requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
jinja2==3.1.6
bleach==6.2.0
pymysql==1.1.1
email-validator==2.2.0
httpx==0.28.1
bcrypt==4.0.1
